+++ servant-reimagined-server/src/Servant/Reimagined/Server.hs
@@ -148,6 +148,7 @@ module Servant.Reimagined.Server
   , ClientStreamHandler
   , BidiStreamHandler
   , sseResponse
+  , sseResponseSync
   , sseChunk
   ) where